refactor: Fare, DiscountPolicy 클래스 테스트 코드 구현 클래스 별로 분리
